Add tool for encrypting/decrypting
