+++ t/data/authentication.conf
@@ -405,6 +405,20 @@ type=AD
 host=127.0.0.1
 cache_match=0
 
+[LDAPWITHENCRYPTEDPASSWORD]
+description=pf-test
+password=
+scope=sub
+binddn=CN=test,DC=inverse,DC=ca
+basedn=DC=ldap,DC=inverse,DC=ca
+usernameattribute=user
+connection_timeout=5
+encryption=none
+port=33389
+type=AD
+host=127.0.0.1
+cache_match=1
+
 [htpasswd-stripped]
 realms=
 set_access_level_action=
